endpointFuncs/izifood/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core
azure-cosmos
azure-functions
certifi
charset-normalizer
idna
requests
six
typing-extensions
urllib3
